pyproject.toml
[project]
name = "arxiv-classifier-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "datasets>=4.4.2",
    "jupyter>=1.1.1",
    "kaggle>=1.7.4.5",
    "matplotlib>=3.9.4",
    "numpy<2.0",
    "pandas>=2.3.3",
    "pydantic-settings>=2.11.0",
    "seaborn>=0.13.2",
    "tensorflow-macos==2.16.2",
    "tensorflow-metal==1.1.0",
    "tf-keras==2.16.0",
    "tqdm>=4.67.1",
    "transformers>=4.57.3",
    "wandb>=0.24.0",
    "tf2onnx>=1.16.0",
    "django>=5.0",
    "djangorestframework>=3.15",
    "django-cors-headers>=4.3",
    "onnxruntime>=1.18",
    "pytest>=8.0",
    "pytest-timeout>=2.3",
    "scikit-learn>=1.4",
    "gradio>=4.0",
    "gunicorn>=21.0",
    "pyyaml>=6.0",
    "nltk>=3.9.2",
]
